Revert "[nrf fromtree] Bluetooth: Host: Set scan option type to uint8_t"
This reverts commit 886c21075774647816bf2a311a95799d8d45c4cf.

+++ include/zephyr/bluetooth/bluetooth.h
@@ -2065,7 +2065,7 @@ struct bt_le_scan_param {
 	uint8_t  type;
 
 	/** Bit-field of scanning options. */
-	uint8_t options;
+	uint32_t options;
 
 	/** Scan interval (N * 0.625 ms).
 	 *
